worker/storageprovisioner: split volume.go
